==> scripts/00_behavior_wrangle.Rmd <==
---
title: "Behavior data wrangle"
output: md_document
---

```{r setup, message=F}
## load libraries 
library(tidyverse) ## for respahing data

## set output file for figures 
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# because this csv file has some rows with text, I have to specific which columns should be double
mycols <- cols(
  "Year"  = col_integer(),  "ID"  = col_character(), "Genotype" = col_character(), "TrainProtocol"  = col_character(), 
"TrainSequence"  = col_factor(),     "TrainGroup"  = col_factor(),        "Day"   = col_integer(),  "TrainSession"   = col_factor(),    
"ShockOnOff"  = col_character(),       "PairedPartner" = col_character(),     "Experimenter"  = col_character(),     "Housing"  = col_character(),         
 "TestLocation" = col_character(),      "filename" = col_character(),          "p-miss" = col_double(),            "TotalTime(s)"  = col_double(),     
"TotalPath(Arena)" = col_double(),   "Speed (Arena)" = col_double(),      "sd Speed(Arena)" = col_double(),    "Linearity(Arena)"  = col_double(), 
 "TotalPath(m)"= col_double(),        "#Entrances"  = col_double(),        "Time1stEntr" = col_double(),        "Path1stEntr"  = col_double(),      
 "Speed1stEntr(cm/s)" = col_double(), "Entr/Dist(1/m)" = col_double(),     "#Shock"  = col_double(),            "Time1stShock"  = col_double(),     
 "Path1stShock"  = col_double(),      "Speed (cm/s)"   = col_double(),     "SdSpeed"  = col_double(),           "Linearity" = col_double(),         
 "MaxTimeAvoid" = col_double(),       "MaxPathAvoid"  = col_double(),      "Time2ndEntr"   = col_double(),      "Path2ndEntr" = col_double(),       
"Speed2ndEntr"   = col_double(),     "TimeTarget" = col_double(),         "pTimeTarget"  = col_double(),       "pTimeCCW" = col_double(),          
 "pTimeOPP"    = col_double(),        "pTimeCW" = col_double(),            "RayleigLength" = col_double(),      "RayleigAngle"   = col_double(),    
 "PolarAvgVal"  = col_double(),       "PolarSdVal"= col_double(),          "PolarMinVal"  = col_double(),       "PolarMinBin"   = col_double(),     
 "Min50%RngLoBin" = col_double(),     "Min50%RngHiBin" = col_double(),     "PolarMaxVal"  = col_double(),       "PolarMaxBin" = col_double(),       
 "Max50%RngLoBin" = col_double(),     "Max50%RngHiBin" = col_double(),     "AnnularMinVal"    = col_double(),   "AnnularMinBin"  = col_double(),    
 "AnnularMaxVal"  = col_double(),     "AnnularMaxBin" = col_double(),      "AnnularAvg"  = col_double(),        "AnnularSd"  = col_double(),        
 "AnnularSkewnes"  = col_double(),    "AnnularKurtosis"= col_double() 
)

# import raw data
behav <- read_csv("../data/00_Data2013_2016_forAnalysis.csv", col_types = mycols) 
behav <- as.data.frame(behav)
names(behav)

# remove some columns that are not behavioral measures
columnstoremove <- c("PolarAvgVal" , "PolarSdVal", "PolarMinVal" ,
                     "PolarMinBin", "Max50%RngLoBin", "Min50%RngHiBin" , 
                     "PolarMaxVal", "PolarMaxBin", "Max50%RngLoBin", "Max50%RngHiBin",
                     "p-miss", "TotalTime(s)", "TotalPath(m)",  "Time1stShock",  
                     "Path1stShock", "Speed (cm/s)", "SdSpeed", "Linearity",
                     "AnnularMinVal" , "AnnularMinBin" , "AnnularMaxVal" , 
                     "AnnularMaxBin", "AnnularAvg"  , "AnnularSd" )

# keep only 2015 animals and remove unnecessary columns
behav <- behav %>% dplyr::filter(Year == "2015") %>% 
                   dplyr::select(-columnstoremove) %>% 
                  droplevels()

# rename some columns
names(behav)[names(behav)=="#Shock"] <- "NumShock"
names(behav)[names(behav)=="#Entrances"] <- "NumEntrances"
names(behav)[names(behav)=="pTimeTarget"] <- "pTimeShockZone"
names(behav)[names(behav)=="TimeTarget"] <- "TimeShockZone"
names(behav)[names(behav)=="Speed (Arena)"] <- "Speed"
names(behav)[names(behav)=="TotalPath(Arena)"] <- "TotalPath"
names(behav)[names(behav)=="Speed1stEntr(cm/s)"] <- "Speed1stEntr"
names(behav)[names(behav)=="Speed2ndEntr(cm/s)"] <- "Speed2ndEntr"
names(behav)[names(behav)=="Linearity(Arena)"] <- "Linearity"
names(behav)[names(behav)=="sd Speed(Arena)"] <- "sdSpeed"
names(behav)[names(behav)=="Entr/Dist(1/m)"] <- "EntrPerDist"


# check which measures are included
measures <- behav %>% select(TotalPath:AnnularKurtosis)
names(measures)

# make new factors
behav <- behav %>%
  mutate(APA = paste(TrainSequence, TrainGroup, sep = "-")) %>%
  mutate(treatment = fct_recode(APA,
                                "standard.yoked" = "train-train-yoked",
                                "standard.trained" = "train-train-trained",
                                "conflict.yoked" = "train-conflict-yoked",
                                "conflict.trained" = "train-conflict-trained")) %>%
  mutate(training = fct_collapse(treatment,
                                      trained = c("standard.trained", "conflict.trained"),
                                      yoked = c("standard.yoked", "conflict.yoked")))  %>%
  mutate(trial = fct_collapse(TrainSession,
                                      "T4_C1" = c("T4", "C1"),
                                      "T5_C2" = c("T5", "C2"),
                                      "T6_C3" = c("T6", "C3"),
                                      "Hab"  = "Hab",
                                      "T1" = "T1",
                                      "T2"= "T2",
                                      "T3"= "T3",
                                      "Retest" = "Retest",
                                      "Retention" = "Retention"))

behav$trial <- factor(behav$trial, 
                                  levels = c("Hab", "T1","T2","T3","Retest",
                                   "T4_C1","T5_C2","T6_C3","Retention"))
behav$trialNum <- as.numeric(behav$trial)

behav <- behav %>%  dplyr::select(ID,Day,treatment, training,trial,
                                 trialNum, ShockOnOff, PairedPartner,
                                 TotalPath:AnnularKurtosis) %>%
                    dplyr::arrange(ID)

head(behav)


## Fix the NumShocks problem but it doesn't know the the yoked got shocked when the trained got shocked rather than they enterer the shock zone.
# Steps:
# 1. Create a "shocks1" df that have only ID, PairedPartner, training, trialNum, NumShock, ShockOnOff
# 2. Separate the tarined and yoked mice into two dataframes, relabel columsn so the two dfs can be rejoined
# 3. Creat a "shocks2" df by join the two dataframes and replace all NumShock value for all the yoked animals, and set it to 0 when shock is off
# 4. Sparate the trained and yoked again and rename so they can be join into a final "shocks3" datafram by 
# 5. Join the behavior data with the new shocks data. Replace old NumShock with new NumShock

shocks1 <- behav %>% select(ID, PairedPartner, training, trialNum, NumShock, ShockOnOff) 

trained <- shocks1 %>% filter(training == "trained") %>% 
  rename(trained = ID, yoked = PairedPartner, ttraining = training,  tNumShock = NumShock)
yoked <- shocks1 %>% filter(training == "yoked") %>% 
  rename(yoked = ID, trained = PairedPartner, ytraining = training, yNumShock = NumShock)

shocks2 <- full_join(trained,yoked) %>% select(trained, yoked, ShockOnOff, trialNum, tNumShock) 
shocks2$tNumShock <- ifelse(shocks2$ShockOnOff == "Off", 0, shocks2$tNumShock )

trained <- shocks2 %>% select(trained, trialNum, tNumShock) %>% rename(ID = trained )
yoked <- shocks2 %>% select(yoked, trialNum, tNumShock) %>% rename(ID = yoked )
shocks3 <- rbind(trained, yoked)

behav <- full_join(behav, shocks3) %>% mutate(NumShock = tNumShock) %>% select(-tNumShock)

# add new measure 
behav <- behav %>%  mutate(ShockPerEntrance = NumShock / NumEntrances) %>%  
  mutate(ShockPerEntrance = ifelse(is.na(ShockPerEntrance), 0, ShockPerEntrance))

```



```{r}
write.csv(behav, "../data/00_behaviordata.csv", row.names = F)
```
