==> scripts/04_correlations.Rmd <==
---
title: "Correlation Analysis"
output: md_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

knitr::opts_chunk$set(fig.path = '../figures/04_correlations/')
```


```{r packages}
library(tidyverse)
library(corrr)
library(cowplot)
library(Hmisc) # for correlations with pvalue
library(ggrepel)


source("./figureoptions.R")
source("./functions_RNAseq.R")
```

## Sample information and PC1

```{r DGDEGs}
# read the sample data, set levels, join iwth behvior PCA data
colData <- read.csv("../data/00_colData.csv", row.names = 1, stringsAsFactors = T) 
pca.Rn <- read_csv("../data/suppltable-3.csv") %>% dplyr::filter(trialNum == 9)

colData <- left_join(colData, pca.Rn)
head(colData)
```

```{r allvsd}
# read all count data prep to join with sample colData

combinePCvsd <- function(filename, whichsubfield, whichgenes){
  
  
  vsd <- read.csv(filename, row.names = 1, check.names = F) 
  vsd$gene <- row.names(vsd)
  vsd$gene <- str_to_title(vsd$gene)
  vsd <- as.data.frame(vsd)
  row.names(vsd) <- vsd$gene
  vsd$gene <- NULL
  vsd <- as.data.frame(t(vsd))
  vsd$sample <- row.names(vsd)
  vsd$mouse <- sapply(strsplit(as.character(vsd$sample),"\\-"), "[", 1)
  vsd$ID <- paste(15, vsd$mouse, sep = "")
  
  vsd <- left_join(colData, vsd) %>%
    select(ID:training, 
           PC1, PC2,
           whichgenes) %>%
    filter(subfield == whichsubfield) 
  print(head(vsd))
  return(vsd)
}

candidategenes <- c("Prkcz", "Prkci", "Wwc1", "Grin1", "Gria1",
         "Pick1", "Nsf", "Fmr1", "Camk2a",
         "Fos", "Fosl2", "Npas4", "Arc")

vsdDG <- combinePCvsd("../data/03_DG_vsdtraining.csv", "DG", candidategenes)
vsdCA1 <- combinePCvsd("../data/03_CA3_vsdtraining.csv", "CA1", candidategenes)
vsdCA3 <- combinePCvsd("../data/03_DG_vsdtraining.csv", "CA3", candidategenes)

# save only genes in all
DGgenes <- names(vsdDG)
CA3genes <- names(vsdCA3)
CA1genes <- names(vsdCA1)
savecols <- intersect(DGgenes, CA3genes)
savecols <- intersect(savecols, CA1genes)


vsdDG <- vsdDG %>% dplyr::select(one_of(savecols)) 
vsdCA1 <- vsdCA1 %>% dplyr::select(one_of(savecols)) 
vsdCA3 <- vsdCA3 %>% dplyr::select(one_of(savecols)) 

allvsd <- rbind(vsdDG, vsdCA1)
allvsd <- rbind(allvsd, vsdCA3)
head(allvsd)

write.csv(allvsd, "../data/for3Dplot.csv")
```





```{r Arc}
summary(lm( PC1 ~ Arc, data = vsdDG))
summary(lm( PC2 ~ Arc, data = vsdDG))

cor.test(vsdDG$PC1, vsdDG$Arc, method = c("pearson"))
cor.test(vsdDG$PC2, vsdDG$Arc, method = c("pearson"))

iconDG <- png::readPNG("../figures/00_schematics/DG.png")
iconDG <-  grid::rasterGrob(iconDG, interpolate=TRUE)

vsdDG$training <- factor(vsdDG$training, levels = levelstraining)

a <- ggplot(vsdDG, aes(x = Arc, y = PC1)) +
   geom_point(aes( color = training)) + 
   geom_smooth(method = "lm", color = "grey") +
   scale_color_manual(values = allcolors) +
  theme_ms() +
   theme(legend.position = "bottom",
         axis.title.x = element_text(face = "italic"),
         legend.title = element_blank(), 
         legend.key.size = unit(0.25, "cm"))  +
  labs(subtitle = "r = 0.81, p = 0.0002") +
   annotation_custom(iconDG, ymin = 6, ymax = 11, xmin = 7.5, xmax = 9)
a
```
