==> scripts/06_searchfavgenes.Rmd <==
---
title: "Search favorite gene Analysis"
output: md_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.path = '../figures/06_favegenes/', cache = T, cache.lazy = F)
```

```{r packages}
suppressMessages(library(tidyverse))
suppressMessages(library(cowplot))
suppressMessages(library(corrr)) # for easy correlations
suppressMessages(library(Hmisc)) # for correlations with pvalue

source("./figureoptions.R")
source("./functions_RNAseq.R")
```

## Sample information and PC1

```{r vsds}
# meta data
colData <- read_csv("../data/00_colData.csv")

geneids <- read_csv("../data/00_geneids.csv") %>% distinct(gene) %>% mutate(gene = toupper(gene))

# pca data
pca.Rn <- read_csv("../data/01_pca.all.csv") %>% 
  filter(trialNum == 9) %>%  
  select(ID:PC2) %>%
  left_join(colData) %>% drop_na() %>% select(ID, treatment, training, trialNum, Day, RNAseqID, subfield, PC1, PC2)
head(pca.Rn)

prepvsdforjoin <- function(pathtovsd, mysubfield){
  #read all count data
  vsd <- read_csv(pathtovsd) 
  vsd$gene <- vsd$X1
  vsd$gene <- str_to_title(vsd$gene)
  vsd <- vsd %>% arrange(gene) %>%  select(gene,everything())
  vsd <- as.data.frame(vsd)
  vsd$X1 <- NULL
  row.names(vsd) <- vsd$gene
  vsd$gene <- NULL
  vsd <- as.data.frame(t(vsd))

  vsd$RNAseqID <- row.names(vsd)
  vsd <- vsd %>%  select(RNAseqID, everything())
  vsd <- left_join(pca.Rn, vsd)  %>% drop_na()

  vsd$subfield <- factor(vsd$subfield, levels = levelssubfield)
  vsd$treatment <- factor(vsd$treatment, levels = levelstreatment)
  vsd$training <- factor(vsd$training, levels = levelstraining)
  return(vsd)
}

vsdDG <- prepvsdforjoin("../data/03_DG_vsdtraining.csv")
vsdCA1 <- prepvsdforjoin("../data/03_CA1_vsdtraining.csv")
vsdCA3 <- prepvsdforjoin("../data/03_CA3_vsdtraining.csv")

head(vsdDG)[1:10]
```

## correlation between PC1 and PC2 or Arc



```{r icons}
iconDG <- png::readPNG("../figures/00_schematics/DG.png")
iconDG <-  grid::rasterGrob(iconDG, interpolate=TRUE)

iconCA3 <- png::readPNG("../figures/00_schematics/CA3.png")
iconCA3 <-  grid::rasterGrob(iconCA3, interpolate=TRUE)

iconCA1 <- png::readPNG("../figures/00_schematics/CA1.png")
iconCA1 <-  grid::rasterGrob(iconCA1, interpolate=TRUE)

icontrainedyoked <- png::readPNG("../figures/00_schematics/figure_yokedtrained.png")
icontrainedyoked <-  grid::rasterGrob(icontrainedyoked, interpolate=TRUE)

```


## correlations with PC1

```{r supplefig3, fig.height= 5.5, fig.width=6.69}

# correlations to go with plots a, b, and c
cor.test(vsdDG$PC1, vsdDG$PC2, method = c("pearson"))
cor.test(vsdCA3$PC1, vsdCA3$Arc, method = c("pearson"))
cor.test(vsdCA1$PC1, vsdCA1$Arc, method = c("pearson"))

         
plotcorrelation <- function(df, myx, myy){
  p <- ggplot(df, aes(x = myx, y = myy)) +
   geom_point(aes(color = training)) + 
   geom_smooth(method = "lm", color = "grey") +
   scale_color_manual(values = volcano1) +
  theme_ms() +
   theme(legend.position = "none") +
    labs(y = NULL, x = NULL)
  return(p)
}

p1 <- plotcorrelation(vsdDG, vsdDG$PC1, vsdDG$Arc) + 
  labs(x = "PC1", y = "Arc", subtitle = "r = 0.81, p = 0.0002")  + 
  theme(axis.title.y = element_text(face = "italic")) +
  annotation_custom(iconDG, ymin = 10.5, ymax = 11.8, xmin = -3, xmax = 2)


inset <- plotcorrelation(vsdDG, vsdDG$PC1, vsdDG$PC2) + 
  labs(x = "PC1", y = "PC2") 
inset

g2 = ggplotGrob(inset)
p1g <- p1 + annotation_custom(grob = g2, xmin=1.5, xmax=7.5, ymin=7.25, ymax=9.25)
p1g


p2 <- plotcorrelation(vsdCA3, vsdCA3$PC1, vsdCA3$Arc) + 
  labs(x = "PC1", y = "Arc", subtitle = "r = -0.30, p = 0.327")  + 
  theme(axis.title.y = element_text(face = "italic"),
        legend.position = "bottom",
        legend.title = element_blank(),
        legend.key.size =  unit(0.25, "cm")) +
  annotation_custom(iconCA3, ymin = 9.3, ymax = 9.7, xmin = 2.5, xmax = 7)

p3 <- plotcorrelation(vsdCA1, vsdCA1$PC1, vsdCA1$Arc)  + 
  labs(x = "PC1", y = "Arc", subtitle = "r = 0.20, p = 0.48") + 
  theme(axis.title.y = element_text(face = "italic")) +
    annotation_custom(iconCA1, ymin = 10.5, ymax = 11.3, xmin = 2.5, xmax = 7) 

cor.test(vsdDG$PC1, vsdDG$Igf2, method = c("pearson"))
cor.test(vsdCA3$PC1, vsdCA3$Igf2, method = c("pearson"))
cor.test(vsdCA1$PC1, vsdCA1$Igf2, method = c("pearson"))


p4 <- plotcorrelation(vsdDG, vsdDG$PC1, vsdDG$Igf2) + 
  labs(x = "PC1", y = "Igf2", subtitle = "r = 0.38, p = 0.150")  + 
  theme(axis.title.y = element_text(face = "italic"),
        legend.title = element_blank(),
        legend.key.size =  unit(0.25, "cm")) +
  annotation_custom(iconDG, ymin = 5, ymax = 5.6, xmin = 2.5, xmax = 7)


p5 <- plotcorrelation(vsdCA3, vsdCA3$PC1, vsdCA3$Igf2) + 
  labs(x = "PC1", y = "Igf2", subtitle = "r = 0.48, p = 0.092")  + 
  theme(axis.title.y = element_text(face = "italic"),
        legend.title = element_blank(),
        legend.key.size =  unit(0.25, "cm")) +
  annotation_custom(iconCA3, ymin = 5, ymax = 7, xmin = 2.5, xmax = 7)

p6 <- plotcorrelation(vsdCA1, vsdCA1$PC1, vsdCA1$Igf2)  + 
  labs(x = "PC1", y = "Igf2", subtitle = "r = 0.58, p = 0.024") + 
  theme(axis.title.y = element_text(face = "italic")) +
    annotation_custom(iconCA1, ymin = 6, ymax = 6.3, xmin = 2.5, xmax = 7) 

mylegend = get_legend(p2)

p1p2p3 <- plot_grid(p1g,p2 + theme(legend.position = "none"), p3, 
                    p4,p5,p6,
                    labels = "auto", label_size = 8, nrow = 2)

supplfig3 <- plot_grid(p1p2p3, mylegend, nrow = 2, rel_heights = c(1,0.1))
supplfig3

pdf(file="../figures/06_favegenes/supplfig3.pdf", width=6.69, height=5.5)
plot(supplfig3)
dev.off()

pdf(file="../figures/supplfig3.pdf", width=6.69, height=5.5)
plot(supplfig3)
dev.off()

```


## Correlate ALL genes with PC1 and PC2



## Candidate gene

```{r allcandidates}
classicmemgenes <- c("Camk2a" , "Fmr1", "Gria2",  "Igf2" , "Mtor", "Nsf" , 
                     "Pick1", "Prkcb" ,  "Prkci",  "Prkcz", "Wwc1")

stabilizationgenes  <- c("Igf2bp2", "Lama1"  , "Lamb1"  , "Lamc1" ,  "Tnc" ,
                         "Tnxb"  ,  "Tnr" ,    "Gabra1" , "Ptprs"  , "Pnn"  , "Egfr")

ACTINngenes <- c("Limk1","Cfl1", "Rock2")

astrocyticgenes <- c("Aldh1a1", "Aldh1l1" ,"Aldh1l2", "Slc1a2" , "Gfap" ,
                     "Gjb6" ,   "Fgfr3" ,  "Aqp4" ,   "Aldoc")

allcandidates <- c(classicmemgenes, ACTINngenes, stabilizationgenes, astrocyticgenes) 
```

```{r makecorrrmat}
# Use `tail(names(vsd),10)` to see that last genes is "Zzz3"
makecorrrmatrix <- function(df){
  forcorall <-  df %>% select(PC1, PC2, allcandidates)
  corrrmat <- correlate(forcorall) 
  return(corrrmat)
}
```

```{r makecorrrmat1}
corrrDG <- makecorrrmatrix(vsdDG)     
```

```{r makecorrrmat2}
corrrCA3 <- makecorrrmatrix(vsdCA3)      
```

```{r makecorrrmat3}
corrrCA1 <- makecorrrmatrix(vsdCA1)    
```

```{r classicmemgenes}
plotcorrrs2 <- function(favgenes, corrs, whichsubfield, myPC1label, matrixlabel , mylabels){
  
  df <- corrs %>% 
    focus(PC1, PC2, favgenes,  mirror = TRUE)  %>% 
    arrange(desc(PC1))
  
  p1 <- df %>% 
    focus(PC1) %>%
    dplyr::filter(rowname != "PC2") %>%
    ggplot(aes(x = reorder(rowname, desc(rowname)), PC1, fill = PC1, label = round(PC1,2))) +
     geom_col() + coord_flip() +
    scale_fill_gradient2(low = "#67a9cf",  high = "#ef8a62", midpoint = 0) +
    theme_ms() +
    theme(legend.position = "none", axis.text.y = element_text(face = "italic")) +
    ylim(-0.75, 0.75) +
    labs(x = whichsubfield, y = myPC1label) +
     geom_hline(yintercept = -0.6, linetype = "dashed", color = "grey", size = 0.5) +
     geom_hline(yintercept = 0.6, linetype = "dashed", color = "grey", size = 0.5)

  dflong <- df %>%
    pivot_longer(-rowname, names_to = "colname") %>%
    dplyr::filter(rowname != "PC1", rowname != "PC2")%>%
    dplyr::filter(colname != "PC1", colname != "PC2")

  p3 <-  ggplot(dflong, aes(x = rowname, y = reorder(colname, desc(colname)), 
                            fill= value, label = round(value,2))) + 
    geom_tile() +
    scale_fill_gradient2(low = "#67a9cf",mid = "white", 
                         high = "#ef8a62", na.value = "white",
                         limits = c(-1,1)) +
    theme_classic(base_size = 8) +
    labs(x = matrixlabel, y = " ") +
    theme(axis.text = element_text(face = "italic"),
          legend.position = "none") +
    geom_text(size=2.5, aes(label= ifelse(value > 0.6, round(value,2),
                                          ifelse(value < -0.6, round(value,2), " "))))
  
  
  p123 <- plot_grid(p1, p3,  
                   nrow = 1, labels = mylabels, label_size = 8, rel_widths = c(0.9,1.5))
  
  #filename <- paste("../data/06_", whichsubfield, "_candidatecorrelations.csv", sep = "")
  #write.csv(df, filename)
  return(p123)
  
}

p4 <- plotcorrrs2(classicmemgenes, corrrDG, "DG", NULL, NULL,  c("a", "b" ))
p5 <- plotcorrrs2(classicmemgenes, corrrCA3, "CA3", NULL, NULL, c("c","d"))
p6 <- plotcorrrs2(classicmemgenes, corrrCA1, "CA1", 
                  "Correlation to PC1 \n (avoidance estimate)",
                  "Candidate genes \n ",
                  c("e", "f"))

p456 <- plot_grid(p4,p5,p6, nrow = 3, rel_heights = c(1,1,1.2))

circuit <- png::readPNG("../figures/00_schematics/figure_hippocircuit.png")
circuit <- ggdraw() +  draw_image(circuit, scale = 1)

p456circuit <- plot_grid(circuit, p456, nrow = 1, rel_widths = c(0.1,1))
p456circuit

pdf(file="../figures/06_favegenes/classicmemgenes.pdf", width=6.69, height=6)
plot(p456circuit)    
dev.off()

pdf(file="../figures/figure_4.pdf", width=6.69, height=6)
plot(p456circuit)    
dev.off()
```


```{r astrocyticgenes}
p7 <- plotcorrrs2(astrocyticgenes, corrrDG, "DG", NULL, NULL, c("a", "b"))
p8 <- plotcorrrs2(astrocyticgenes, corrrCA3, "CA3", NULL, NULL, c("c", "d"))
p9 <- plotcorrrs2(astrocyticgenes, corrrCA1, "CA1", 
                  "Correlation to PC1 \n (avoidance estimate)",
                  "Candidate genes \n",
                  c("e", "f"))

p789 <- plot_grid(p7,p8,p9, nrow = 3, rel_heights = c(1,1,1.2))

p789circuit <- plot_grid(circuit, p789, nrow = 1, rel_widths = c(0.1,1))
p789circuit

pdf(file="../figures/06_favegenes/astrocyticgenes.pdf", width=6.69, height=6)
plot(p789circuit)    
dev.off()

pdf(file="../figures/supplfig4.pdf", width=6.69, height=6)
plot(p789circuit)    
dev.off()

```


## correlations with p-values


```{r Hmisc}
getcandidategenecorrelations <- function(df, candidategenes, whichsubfield){
  
  x <- df %>%
    select(ID, PC1, PC2, candidategenes)
  x <- as.data.frame(x)
  row.names(x) <- x$ID
  x$ID <- NULL
  res <- rcorr(as.matrix(x))
  
  cormat <- res$r
  pmat <- res$P
  
  ut <- upper.tri(cormat)
  
  newdf <- data.frame(
    row = rownames(cormat)[row(cormat)[ut]],
    column = rownames(cormat)[col(cormat)[ut]],
    cor  =(cormat)[ut],
    p = pmat[ut]
    )
  
  newdf <- newdf %>% arrange(p) %>%
    mutate(padj = p.adjust(p, method = "fdr", n = length(p)))
  
  filename <- paste("../data/06_", whichsubfield, "_corrswithpvalue.csv", sep = "")
  write.csv(newdf, filename)
  
  print(head(newdf, 10))
  return(newdf)
  
}

DGcorrswithpvalue <- getcandidategenecorrelations(vsdDG, allcandidates, "DG")
CA3corrswithpvalue <- getcandidategenecorrelations(vsdCA3, allcandidates, "CA3")
CA1corrswithpvalue <- getcandidategenecorrelations(vsdCA1, allcandidates, "CA1")


```
