==> scripts/07_cembrowski.Rmd <==
---
title: "Untitled"
author: "Rayna M Harris"
date: "7/7/2020"
output: html_document
---


## mareker gene analysis

Genes from Cembrowski sublement file 1 <https://elifesciences.org/articles/14997/figures>

```{r markergenes}
library(tidyverse)
# import cembrowski markers
cembrowskisupp <- read.table("../data/elife-14997-supp1-v1.txt", sep="\t", header = T)

# select just columns with gene symbol and enriched column
# then rename gene column and convert to uppercase
cembrowskisupp <- cembrowskisupp %>% dplyr::select(gene_short_name, enriched) 
colnames(cembrowskisupp)[1] <- "gene"
cembrowskisupp$gene <- str_to_upper(cembrowskisupp$gene)
head(cembrowskisupp)

# subset my maker

cembrowksimarkers <- function(subfields){
  mylist <- cembrowskisupp %>% 
    dplyr::filter(enriched %in% subfields) %>% 
    pull(gene)
  return(mylist)
}

levels(cembrowskisupp$enriched)

# dorsal markers 
CA1_markers <- cembrowksimarkers(c("ca1_d", "ca1_d-ca1_v"))
DG_markers <- cembrowksimarkers(c("dg_d", "dg_d-dg_v"))
CA3_markers <- cembrowksimarkers(c("ca3_d",  "ca3_d-ca3_v-ca2-ca1_d-ca1_v"))
```




```{r enrichementvalidation}
#look for markers in each supfield
# make data frames of genes expression results for markers 

marker_summary <- function(mydf, markers){
    df <- mydf %>%
    dplyr::filter(gene %in% c(markers)) 
    #return((df))
    return(summary(df$direction))
}

# comparing all lists with a given brain region
for(i in list(CA1DG, CA1CA3)){
  j <- marker_summary(i, CA1_markers)
  print(i[1, 6])
  print(j)
}

15/(15+3)
15/(15+3)

for(i in list(CA1CA3, CA3DG)){
  j <- marker_summary(i, CA3_markers)
  print(i[1, 6])
  print(j)
}

6/10
7/10

for(i in list(CA1DG, CA3DG)){
  j <- marker_summary(i, DG_markers)
  print(i[1, 6])
  print(j)
}

48/(48+23)
49/(49+22)

```

Then, I checked to see how many of the markers that Cembrowski found to be enriched in discrete dorsal cell populations were also enriched in my comparisons. The enriched coloumn is what percent were confirmed, the depleted column means that the marker was experssed in the opposite direction, and neither means that the Cembrowski marker was not significantly different in expression between the two cell types. Here are the results:

| Maker | Comparison | Enriched | Depleted | Neither |
|-------|------------|----------|----------|---------|
| CA1 | CA1 v DG | 0.83 | 0.00 | 0.17 |
| CA1 | CA1 v CA3 | 0.83 | 0.06 | 0.11 |
| CA3 | CA3 v DG | 0.60 | 0.00 | 0.40 |
| CA3 | CA3 v CA1 | 0.70 | 0.00 | 0.30 |
| DG | DG v CA1 | 0.68 | 0.00 | 0.32 |
| DG | DG v CA3 | 0.69 | 0.00 | 0.31 |

```{r barplot}
mybarplot <- read.csv("../data/02h_markers.csv")
head(mybarplot)

p <- ggplot(data=mybarplot, aes(x=comparison, y=correct, fill = marker)) +
  geom_bar(stat="identity", position=position_dodge()) +
  theme_ms() +
  scale_fill_manual(values = c( "#7570b3","#1b9e77",  "#d95f02"),
                    name = "markers") +
  ylab("% marker genes recovered") +
  xlab("subfields compared") +
  theme(panel.grid.minor=element_blank(),
        panel.grid.major=element_blank(),
        legend.key.width =  unit(0.2, "cm"),
        legend.key.height =  unit(0.1, "cm"),
        axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "bottom") +
  scale_y_continuous(labels = scales::percent,
                     limits = c(0,1)) 
p
```
